+++ src/errors.ts
@@ -22,6 +22,8 @@ export class QueryMutatedState extends NonRetryableError {}
 export class UnknownEffect extends NonRetryableError {}
 export class UnknownCommitAction extends NonRetryableError {}
 export class UnknownDeadLetter extends SolidObjectsError {}
+export class UnknownReminder extends SolidObjectsError {}
+export class ReminderNotPaused extends SolidObjectsError {}
 export class InvalidStreamToken extends SolidObjectsError {}
 export class InvalidComponentToken extends SolidObjectsError {}
 export class UnknownComponent extends SolidObjectsError {}